refactor: moved Utilities.kt from application/shared to domain/shared
