+++ review/src/main/kotlin/br/all/domain/shared/utils/Utilities.kt
@@ -1,4 +1,4 @@
-package br.all.application.shared.utils
+package br.all.domain.shared.utils
 
 fun requireThatExists(doesItExist: Boolean, message: () -> String = { "There is no such element!" }) {
     if (!doesItExist) throw NoSuchElementException(message())
